requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-cloud-aiplatform==1.67.0
vertexai==1.67.0
google-generativeai==0.8.3
toolbox-core==0.3.0
google-adk==0.5.0
httpx==0.27.2
